Fix some failing tests
In most cases the failure is due to committing to a whitespace parser which
prevents us from backtracking and trying a different parse.  In each case the
solution is usually to `try` until we parse something that definitively
indicates that we are on the right track so that we no longer need to
backtrack.